+++ lib/src/signed_video_openssl.c
@@ -108,7 +108,7 @@ openssl_private_key_malloc(sign_or_verify_data_t *sign_data,
   svrc_t status = SV_UNKNOWN_FAILURE;
   SV_TRY()
     // Read private key
-    BIO *bp = BIO_new_mem_buf(private_key, private_key_size);
+    BIO *bp = BIO_new_mem_buf(private_key, (int)private_key_size);
     signing_key = PEM_read_bio_PrivateKey(bp, NULL, NULL, NULL);
     BIO_free(bp);
     SV_THROW_IF(!signing_key, SV_EXTERNAL_ERROR);
@@ -406,7 +406,7 @@ oid_to_type(message_digest_t *self)
     // Point to the first byte of the OID. The |oid_ptr| will increment while decoding.
     encoded_oid_ptr = self->encoded_oid;
     SV_THROW_IF(
-        !d2i_ASN1_OBJECT(&obj, &encoded_oid_ptr, self->encoded_oid_size), SV_EXTERNAL_ERROR);
+        !d2i_ASN1_OBJECT(&obj, &encoded_oid_ptr, (long)self->encoded_oid_size), SV_EXTERNAL_ERROR);
     self->type = EVP_get_digestbyobj(obj);
     self->size = EVP_MD_size(self->type);
   SV_CATCH()
@@ -531,7 +531,7 @@ openssl_encoded_oid_to_str(const unsigned char *encoded_oid, size_t encoded_oid_
   }
 
   // Point to the first byte of the OID. The |oid_ptr| will increment while decoding.
-  if (!d2i_ASN1_OBJECT(&obj, &encoded_oid, encoded_oid_size)) {
+  if (!d2i_ASN1_OBJECT(&obj, &encoded_oid, (long)encoded_oid_size)) {
     goto done;
   }
   OBJ_obj2txt(algo_name, 50, obj, 1);
